Simplify mobile app to location switch + website link
Remove all screens except Login, Register, and a new HomeScreen.
HomeScreen has a single toggle to share GPS location and a button
to open the website. AuthContext now reloads vendor data on startup
using the stored token so the UI is correct after app restarts.
