+++ sunny_sales_mobile/src/context/AuthContext.tsx
@@ -47,16 +47,21 @@ export function AuthProvider({ children }: { children: ReactNode }) {
   // Estado que guarda os dados do vendedor autenticado
   const [vendor, setVendor] = useState<Vendor | null>(null);
 
-  // Carrega o token armazenado ao iniciar a aplicação
+  // Carrega o token armazenado ao iniciar a aplicação e recarrega os dados do vendedor
   useEffect(() => {
     const loadToken = async () => {
       try {
         const storedToken = await AsyncStorage.getItem("token");
         if (storedToken) {
           setToken(storedToken);
+          const res = await api.get("/vendors/me", {
+            headers: { Authorization: `Bearer ${storedToken}` },
+          });
+          setVendor(res.data);
         }
       } catch (error) {
-        console.error("Erro ao carregar token:", error);
+        console.error("Erro ao carregar sessão:", error);
+        await AsyncStorage.removeItem("token");
       }
     };
     loadToken();
