--- sunny_sales_mobile/src/screens/InvoicesScreen.tsx
@@ -1,240 +1,0 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { View, Text, FlatList, StyleSheet, ActivityIndicator, TouchableOpacity, RefreshControl, Linking } from 'react-native';
-import { Ionicons } from '@expo/vector-icons';
-import api, { BASE_URL } from '../services/api';
-import { AuthContext } from '../context/AuthContext';
-
-interface Invoice {
-  id: number;
-  invoice_number: string;
-  issued_at: string;
-  amount: number;
-  status: string;
-  pdf_url: string | null;
-}
-
-export default function InvoicesScreen() {
-  const [invoices, setInvoices] = useState<Invoice[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [refreshing, setRefreshing] = useState(false);
-  const { token, vendor } = useContext(AuthContext);
-
-  const loadInvoices = async () => {
-    if (!vendor) return;
-    try {
-      const res = await api.get(`/vendors/${vendor.id}/invoices`, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-      setInvoices(res.data || []);
-    } catch (error) {
-      console.error('Erro ao carregar faturas:', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    loadInvoices();
-  }, [vendor]);
-
-  const onRefresh = async () => {
-    setRefreshing(true);
-    await loadInvoices();
-    setRefreshing(false);
-  };
-
-  const handleDownloadPDF = (invoice: Invoice) => {
-    if (!invoice.pdf_url) {
-      alert('PDF não disponível para esta fatura');
-      return;
-    }
-    const url = invoice.pdf_url.startsWith('http')
-      ? invoice.pdf_url
-      : `${BASE_URL}${invoice.pdf_url}`;
-    Linking.openURL(url).catch(() => {
-      alert('Não foi possível abrir o PDF');
-    });
-  };
-
-  const formatDate = (dateStr: string) => {
-    return new Date(dateStr).toLocaleDateString('pt-PT', {
-      day: '2-digit',
-      month: '2-digit',
-      year: 'numeric',
-    });
-  };
-
-  const getStatusColor = (status: string) => {
-    switch (status?.toLowerCase()) {
-      case 'paid':
-        return '#4caf50';
-      case 'pending':
-        return '#ff9800';
-      case 'overdue':
-        return '#e74c3c';
-      default:
-        return '#999';
-    }
-  };
-
-  const getStatusLabel = (status: string) => {
-    switch (status?.toLowerCase()) {
-      case 'paid':
-        return 'Paga';
-      case 'pending':
-        return 'Pendente';
-      case 'overdue':
-        return 'Vencida';
-      default:
-        return status;
-    }
-  };
-
-  const renderInvoice = ({ item }: { item: Invoice }) => (
-    <View style={styles.invoiceCard}>
-      <View style={styles.invoiceHeader}>
-        <View style={styles.invoiceIcon}>
-          <Ionicons name="document-text" size={20} color="#1976d2" />
-        </View>
-        <View style={styles.invoiceInfo}>
-          <Text style={styles.invoiceNumber}>{item.invoice_number}</Text>
-          <Text style={styles.invoiceDate}>{formatDate(item.issued_at)}</Text>
-        </View>
-        <View style={styles.invoiceAmount}>
-          <Text style={styles.amount}>{item.amount.toFixed(2)}€</Text>
-          <View style={[styles.statusBadge, { backgroundColor: `${getStatusColor(item.status)}20` }]}>
-            <Text style={[styles.statusText, { color: getStatusColor(item.status) }]}>
-              {getStatusLabel(item.status)}
-            </Text>
-          </View>
-        </View>
-      </View>
-      {item.pdf_url && (
-        <TouchableOpacity
-          style={styles.downloadButton}
-          onPress={() => handleDownloadPDF(item)}
-        >
-          <Ionicons name="download" size={18} color="#1976d2" />
-          <Text style={styles.downloadText}>Descarregar PDF</Text>
-        </TouchableOpacity>
-      )}
-    </View>
-  );
-
-  if (loading) {
-    return (
-      <View style={styles.centerContainer}>
-        <ActivityIndicator size="large" color="#1976d2" />
-      </View>
-    );
-  }
-
-  return (
-    <View style={styles.container}>
-      {invoices.length === 0 ? (
-        <View style={styles.centerContainer}>
-          <Ionicons name="document-outline" size={48} color="#ccc" />
-          <Text style={styles.emptyText}>Nenhuma fatura registada</Text>
-        </View>
-      ) : (
-        <FlatList
-          data={invoices}
-          renderItem={renderInvoice}
-          keyExtractor={(item) => item.id.toString()}
-          contentContainerStyle={styles.listContent}
-          refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} />}
-        />
-      )}
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f5f5f5',
-  },
-  centerContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  listContent: {
-    padding: 16,
-  },
-  invoiceCard: {
-    backgroundColor: '#fff',
-    borderRadius: 8,
-    padding: 16,
-    marginBottom: 12,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  invoiceHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  invoiceIcon: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    backgroundColor: '#e3f2fd',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginRight: 12,
-  },
-  invoiceInfo: {
-    flex: 1,
-  },
-  invoiceNumber: {
-    fontSize: 14,
-    fontWeight: '600',
-    color: '#333',
-    marginBottom: 4,
-  },
-  invoiceDate: {
-    fontSize: 12,
-    color: '#999',
-  },
-  invoiceAmount: {
-    alignItems: 'flex-end',
-  },
-  amount: {
-    fontSize: 16,
-    fontWeight: '700',
-    color: '#1976d2',
-    marginBottom: 4,
-  },
-  statusBadge: {
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-    borderRadius: 12,
-  },
-  statusText: {
-    fontSize: 11,
-    fontWeight: '600',
-  },
-  downloadButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    marginTop: 12,
-    paddingVertical: 8,
-    borderTopWidth: 1,
-    borderTopColor: '#eee',
-    gap: 6,
-  },
-  downloadText: {
-    fontSize: 12,
-    fontWeight: '500',
-    color: '#1976d2',
-  },
-  emptyText: {
-    fontSize: 16,
-    color: '#999',
-    marginTop: 12,
-  },
-});
