--- sunny_sales_mobile/src/screens/SessionsScreen.tsx
@@ -1,174 +1,0 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { View, Text, FlatList, StyleSheet, ActivityIndicator, RefreshControl } from 'react-native';
-import { Ionicons } from '@expo/vector-icons';
-import api from '../services/api';
-import { AuthContext } from '../context/AuthContext';
-
-interface Session {
-  id: number;
-  started_at: string;
-  stopped_at: string | null;
-  is_active: boolean;
-}
-
-export default function SessionsScreen() {
-  const [sessions, setSessions] = useState<Session[]>([]);
-  const [loading, setLoading] = useState(true);
-  const [refreshing, setRefreshing] = useState(false);
-  const { token, vendor } = useContext(AuthContext);
-
-  const loadSessions = async () => {
-    if (!vendor) return;
-    try {
-      const res = await api.get(`/vendors/${vendor.id}/sessions`, {
-        headers: { Authorization: `Bearer ${token}` },
-      });
-      setSessions(res.data || []);
-    } catch (error) {
-      console.error('Erro ao carregar sessões:', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    loadSessions();
-  }, [vendor]);
-
-  const onRefresh = async () => {
-    setRefreshing(true);
-    await loadSessions();
-    setRefreshing(false);
-  };
-
-  const formatDate = (dateStr: string) => {
-    return new Date(dateStr).toLocaleDateString('pt-PT', {
-      year: 'numeric',
-      month: '2-digit',
-      day: '2-digit',
-      hour: '2-digit',
-      minute: '2-digit',
-    });
-  };
-
-  const calculateDuration = (started: string, stopped: string | null) => {
-    const start = new Date(started).getTime();
-    const end = stopped ? new Date(stopped).getTime() : Date.now();
-    const minutes = Math.floor((end - start) / 60000);
-    const hours = Math.floor(minutes / 60);
-    const mins = minutes % 60;
-    return `${hours}h ${mins}m`;
-  };
-
-  const renderSession = ({ item }: { item: Session }) => (
-    <View style={styles.sessionCard}>
-      <View style={styles.sessionHeader}>
-        <View style={[styles.sessionIcon, { backgroundColor: item.is_active ? '#e8f5e9' : '#f5f5f5' }]}>
-          <Ionicons name="clock" size={20} color={item.is_active ? '#4caf50' : '#999'} />
-        </View>
-        <View style={styles.sessionInfo}>
-          <Text style={styles.sessionDate}>{formatDate(item.started_at)}</Text>
-          <Text style={[styles.sessionDuration, { color: item.is_active ? '#4caf50' : '#666' }]}>
-            {calculateDuration(item.started_at, item.stopped_at)}
-          </Text>
-        </View>
-        <View style={[styles.statusBadge, { backgroundColor: item.is_active ? '#e8f5e9' : '#f0f0f0' }]}>
-          <Text style={[styles.statusText, { color: item.is_active ? '#4caf50' : '#999' }]}>
-            {item.is_active ? 'Ativa' : 'Encerrada'}
-          </Text>
-        </View>
-      </View>
-    </View>
-  );
-
-  if (loading) {
-    return (
-      <View style={styles.centerContainer}>
-        <ActivityIndicator size="large" color="#1976d2" />
-      </View>
-    );
-  }
-
-  return (
-    <View style={styles.container}>
-      {sessions.length === 0 ? (
-        <View style={styles.centerContainer}>
-          <Ionicons name="clock-outline" size={48} color="#ccc" />
-          <Text style={styles.emptyText}>Nenhuma sessão registada</Text>
-        </View>
-      ) : (
-        <FlatList
-          data={sessions}
-          renderItem={renderSession}
-          keyExtractor={(item) => item.id.toString()}
-          contentContainerStyle={styles.listContent}
-          refreshControl={<RefreshControl refreshing={refreshing} onRefresh={onRefresh} />}
-        />
-      )}
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#f5f5f5',
-  },
-  centerContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  listContent: {
-    padding: 16,
-  },
-  sessionCard: {
-    backgroundColor: '#fff',
-    borderRadius: 8,
-    padding: 16,
-    marginBottom: 12,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 3,
-  },
-  sessionHeader: {
-    flexDirection: 'row',
-    alignItems: 'center',
-  },
-  sessionIcon: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginRight: 12,
-  },
-  sessionInfo: {
-    flex: 1,
-  },
-  sessionDate: {
-    fontSize: 14,
-    color: '#666',
-    marginBottom: 4,
-  },
-  sessionDuration: {
-    fontSize: 16,
-    fontWeight: '600',
-  },
-  statusBadge: {
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-    borderRadius: 12,
-  },
-  statusText: {
-    fontSize: 12,
-    fontWeight: '500',
-  },
-  emptyText: {
-    fontSize: 16,
-    color: '#999',
-    marginTop: 12,
-  },
-});
